Testserver

+++ src/main/java/gamza/project/gamzaweb/Service/Impl/ProjectServiceImpl.java
@@ -354,7 +354,7 @@ public class ProjectServiceImpl implements ProjectService {
                                 project.getId(),
                                 project.getName(),
                                 project.getApplication().getOuterPort(),
-                                fileUploader.getFileUrl(project)))
+                                fileUploader.recentGetFileUrl(project)))
                 .collect(Collectors.toList());
 
         List<ProjectPerResponseDto> completeProjects = projects.stream()
@@ -364,7 +364,7 @@ public class ProjectServiceImpl implements ProjectService {
                                 project.getId(),
                                 project.getName(),
                                 project.getApplication().getOuterPort(),
-                                fileUploader.getFileUrl(project)))
+                                fileUploader.recentGetFileUrl(project))) // 이부분을 위 아래 수정했음, 가장 최근 프로젝트 link가 반환되도록 함 얘기해볼까 어떻게할지? 일단 오류 수정됨
                 .collect(Collectors.toList());
 
         return ProjectListPerResponseDto.builder()
@@ -526,9 +526,15 @@ public class ProjectServiceImpl implements ProjectService {
 
 
     @Override
+    @Transactional
     public void removeFixedExecutionApplication(HttpServletRequest request, Long id) {
-        userValidate.validateUserRole(request);
-        projectValidate.validateProject(id);
+        userValidate.validateUserRole(request); // 오 ㅋㅋ 이걸 사용하네 ㅇㅈ // 어? 이 validateUserRole 메서드 내가만든거구나 뭐임? 푸핫
+        ProjectEntity project = projectValidate.validateProject(id);
+
+        if (!project.isFixedState()) {
+            throw new BadRequestException("해당 프로젝트는 수정 요청 상태가 아닙니다.", ErrorCode.INTERNAL_SERVER_EXCEPTION);
+        }
+
         projectRepository.deleteById(id);
     }
